+++ tests/Entity/ContextTest.php
@@ -1,0 +1,22 @@
+<?php
+
+/*
+ * This file is part of the Sonata Project package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sonata\ClassificationBundle\Tests\Entity;
+
+use Sonata\ClassificationBundle\Entity\BaseContext;
+
+abstract class ContextTest extends BaseContext
+{
+    public function getId()
+    {
+        return $this->id;
+    }
+}
